pyproject.toml
[project]
name = "foundry-local-voice2action"
version = "0.1.0"
description = "Voice to Action demo using Microsoft Foundry Local and Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-framework",
    "gtts",
    "foundry-local-sdk",
    "openai-whisper",
    "openai",
    "pytest",
]
